libuv/ext: add UV_HANDLE_INTERNAL flags for uv_close.
we need to add this flags to fix uv_loop_close failed issue in workaround ways.

When the application calls sub and unsub multiple times in a loop,
it will cause the handle to be hung on the loop's handle_queue every time uv_poll_init is called.

When uv_loop_close is called, it will be found that there are still many same handles in the loop's
handle queue. The released handle causes loop close to fail.

Therefore, we first wanted to add uv_close to each uv_poll_init to remove the handle from the
handle queue of the Loop, but we found that sub and unsub would process two asynchronous uv_closes
twice in one loop, resulting in a crash.

Therefore, we add a flag to the handle to avoid this detection without causing a memory leak.

Signed-off-by: dongjiuzhu1 <dongjiuzhu1@xiaomi.com>
Signed-off-by: likun17 <likun17@xiaomi.com>